updated cart page updated ticket controller changed data type in database for price in events table

+++ database/migrations/2020_01_08_160841_create_events_table.php
@@ -19,7 +19,7 @@ class CreateEventsTable extends Migration
             $table->text('description');
             $table->bigInteger("user_id");
             $table->bigInteger("places");
-            $table->bigInteger("price");
+            $table->float("price");
             $table->boolean('isFree');
             $table->timestamps();
         });
